feat(webview): share TEC-1G keycap keypad with TEC-1
- Add common/tec-keypad.ts and move layout tables to common/tec-keypad-layout.ts
- TEC-1 uses createTecKeypad without SysCtrl; TEC-1G wraps with status LED wiring
- Move keycap and sysctrl CSS into common/styles.css; keep tec1 breakpoints only
- Align TEC-1 keyboard shortcuts with TEC-1G (arrow mapping)

Made-with: Cursor
