+++ webview/tec1/styles.css
@@ -1,15 +1,7 @@
+/* TEC-1 — layout breakpoints only. Keypad keycaps: ../common/styles.css */
+
 @container (min-width: 440px) {
       .layout {
         grid-template-columns: auto 260px;
       }
-    }
-    .keypad {
-      display: grid;
-      grid-template-columns: 56px repeat(4, 48px);
-      gap: 8px;
-      align-items: center;
-    }
-    .keypad:focus-visible {
-      outline: 2px solid var(--vscode-focusBorder);
-      outline-offset: 2px;
     }
